Flatpak builds 2 electric boogaloo
Github's action vms have weird permissions around namespaces which makes
it not possible to build flatpaks, however the wonderful devs who
maintain flatpak and flathub have a solution, a action and container that build
everything. This however means we have to use a container *without* a
package manager and have to hack around this downloading static binaries
of docker. We shouldn't have to bump the version actively.