Ensure camera configuration works on enumerate_camera_devices
enumerate_camera_devices previously did not check if a camera
configuration worked. This resulted in users being presented with a set
of configurations, some of which did not work. enumerate_camera_devices
now attempted to capture a frame. If this does not work the
configuration is not returned. This greatly reduces the number of
"false" configurations. However there are some configurations which do
not track but are still appended, I am unsure how to test these without
requiring the camera be pointed at a person on startup.